[Fix #285] 메인페이지 에러 처리 완료
