+++ app/(withGlobalLayout)/main/page.tsx
@@ -1,34 +1,32 @@
 "use client";
+import { useEffect, useRef, useState } from "react";
+import Image from "next/image";
 import VisitEmptyImage from "@/assets/images/visit_empty.svg";
 import useConnectStore from "@/store/connect-store";
-import Image from "next/image";
 import S from "@/style/mainpage/main.module.css";
+import { toast } from "react-toastify";
 import { Tables } from "@/types/supabase";
 import GoTopButton from "@/utils/GoTopButton";
 import { socket } from "@/utils/socket/socket";
 import { checkUserLogIn, getUserInfo } from "@/utils/supabase/authAPI";
 import { getRoomsWithKeyword } from "@/utils/supabase/roomAPI";
-import { useRouter } from "next/navigation";
-import { useEffect, useRef, useState } from "react";
-import { toast } from "react-toastify";
 import MainCreateRoom from "@/components/modal/CreateRoomModal";
-import { useCreateStore } from "../../../store/toggle-store";
-import useSocketOn from "@/hooks/useSocketOn";
+import { useCreateStore } from "@/store/toggle-store";
 import MainVisual from "@/components/main/MainVisual";
 import RoomSearch from "@/utils/RoomSearch";
 import RoomListItem from "@/components/main/RoomListItem";
+import useGetRoomsSocket from "@/hooks/useGetRoomsSocket";
 
 const Mainpage = () => {
+  const { rooms, setRooms } = useGetRoomsSocket();
   const { isCreate, setIsCreate } = useCreateStore();
   const { userId, nickname, setRoomId, setUserId, setUserNickname } = useConnectStore();
-  const [rooms, setRooms] = useState([] as Tables<"room_table">[]);
   const [search, setSearch] = useState("");
   const isGoInClick = useRef(false);
-  const roomId = useRef("");
-  const router = useRouter();
 
   useEffect(() => {
     socket.connect();
+    socket.emit("enterMafia", 0, 20);
 
     const checkUserInfo = async () => {
       const userInfo = await getUserInfo();
@@ -41,94 +39,8 @@ const Mainpage = () => {
         // setUserNickname(userInfo.user_metadata.nickname);
       }
     };
-
     checkUserInfo();
   }, []);
-
-  const sockets = {
-    enterMafia: (rooms: Tables<"room_table">[]) => {
-      setRooms(rooms);
-    },
-    joinRoom: () => {
-      if (roomId.current) {
-        router.push(`/room/${roomId.current}/`);
-      }
-    },
-    joinRoomError: (message: string) => {
-      isGoInClick.current = false;
-      toast.error(message);
-      // router.refresh();
-    },
-    fastJoinRoom: (room_id: string) => {
-      router.push(`/room/${room_id}/`);
-      setRoomId(room_id);
-    },
-    fastJoinRoomError: (message: string) => {
-      isGoInClick.current = false;
-      toast.error(message);
-    }
-  };
-
-  useSocketOn(sockets);
-
-  // useEffect(() => {
-  //   //NOTE -  서버와 연결
-  //   socket.connect();
-
-  //   socket.emit("enterMafia", 0, 20);
-
-  //   socket.on("enterMafia", (rooms) => {
-  //     setRooms(rooms);
-  //   });
-  //   socket.on("enterMafiaError", (message) => {
-  //     toast.error(message);
-  //   });
-
-  //   socket.on("joinRoom", () => {
-  //     if (roomId.current) {
-  //       router.push(`/room/${roomId.current}/`);
-  //     }
-  //   });
-
-  //   socket.on("joinRoomError", (message) => {
-  //     isGoInClick.current = false;
-  //     toast.error(message);
-  //     // router.refresh();
-  //   });
-
-  //   socket.on("fastJoinRoom", (room_id) => {
-  //     router.push(`/room/${room_id}/`);
-  //     setRoomId(room_id);
-  //   });
-
-  //   socket.on("fastJoinRoomError", (message) => {
-  //     isGoInClick.current = false;
-  //     toast.error(message);
-  //   });
-
-  //   const checkUserInfo = async () => {
-  //     const userInfo = await getUserInfo();
-
-  //     // 세션 스토리지에 저장
-  //     if (userInfo) {
-  //       setUserId(crypto.randomUUID());
-  //       setUserNickname(crypto.randomUUID());
-  //       // setUserId(userInfo.id);
-  //       // setUserNickname(userInfo.user_metadata.nickname);
-  //     }
-  //   };
-
-  //   checkUserInfo();
-
-  //   return () => {
-  //     socket.off("joinRoom");
-  //     socket.off("joinRoomError");
-  //     socket.off("fastJoinRoom");
-  //     socket.off("fastJoinRoomError");
-  //     socket.off("enterMafia");
-  //     socket.off("enterMafiaError");
-  //   };
-  // }, []);
 
   //NOTE - 로그인 체크, 공통 로직 함수 정의
   const loginErrorHandler = async (emitCallback: () => void) => {
@@ -183,6 +95,9 @@ const Mainpage = () => {
     }
   };
 
+  //NOTE - 방 목록 렌더링
+  if (!rooms) return <div>방 목록 리스트 불러오는 중입니다.</div>;
+
   return (
     <main className={S.main}>
       <section className={S.visualSection}>
@@ -207,7 +122,7 @@ const Mainpage = () => {
               </div>
             </div>
           </div>
-          {rooms.length > 0 ? (
+          {rooms ? (
             <ul className={S.roomList}>
               {rooms.map((item) => (
                 <RoomListItem item={item} joinRoomHandler={joinRoomHandler} />
